backend/pyproject.toml
[project]
name = "datax-backend"
version = "1.0.0"
description = "Data-X: Document-Analytical Platform - Docling-first Architecture"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Data-X Team" }
]
keywords = ["docling", "document-analysis", "data-quality", "fastapi", "mongodb"]

dependencies = [
    # Core Framework
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.9",
    
    # Document Processing (Docling-first)
    "docling>=2.78.0",
    
    # Data Processing
    "pandas>=3.0.0",
    "pandera>=0.18.0",
    "openpyxl>=3.1.0",
    "scipy>=1.11.0",
    
    # Database (PyMongo Async - Sprint 3)
    "pymongo>=4.10.0,<4.17",
    
    # LLM & Embeddings
    "litellm>=1.50.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    
    # Observability & Logging
    "structlog>=24.1.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-arq>=0.24.0",
    "opentelemetry-instrumentation-pymongo>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
    
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Forensics & Security (FraudGuard)
    "pikepdf>=8.0.0",
    "pyhanko>=0.20.0",
    "benford-py>=0.2.0",
    "satcfdi>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
enterprise = [
    "opensearch-py>=2.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
